requirements.txt
Django==2.1.1
django-environ==0.4.5
djangorestframework==3.8.2
psycopg2==2.7.3.2
pytz==2018.5
